Loading env file issue fixed
Apparently we needed to be more specific where the env file was. It also had to be added to an assets folder under app/src/main for it to work. (Also possibly had to be called "env" and not ".env" file according to the guide)

Also added a single empty string inside the AddEventFragment, so that it was correct according to the amount of Event attributes