IT IS ACTUALLY LOADING DATA FROM FIREBASE OMGGG!!!!
Refactored eventDate to be a Long to better comply with the Unix timestamp we have decided to use for the Events.

Also used "Events" as the .getReference in TimelineFragment.

Known Bug: The user can't currently press to go to the login page, to actually login. Which will be looked at after this commit. It just crashes when the user does so (which of course is not ideal)